requirements.txt
assume==0.4.0
atomicwrites==1.4.0
attrs==20.2.0
boto3==1.16.16
botocore==1.19.16
colorama==0.4.4
importlib-metadata==2.0.0
iniconfig==1.1.1
jmespath==0.10.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.8.1
pywin32==228
PyYAML==4.2b4
s3transfer==0.3.3
selenium==3.141.0
six==1.15.0
toml==0.10.2
urllib3==1.25.11
zipp==3.4.0
